dswifi: - add WIFIINIT_OPTION_USEHEAP_96 - force flags so they mix up between u16 and u32 types without issues. - by allocing 96K of ram in dswifi, more sockets can be opened successfully. (otherwise socket allocator will run out of space, if 3 or 4 are open at the same time)

+++ src/common/wifi_shared.h
@@ -34,18 +34,18 @@ SOFTWARE.
 #include "dsregs.h"
 #include "dsregs_asm.h"
 
-#define WIFIINIT_OPTION_USELED		   0x0002
+#define WIFIINIT_OPTION_USELED		   (uint16)(0x0002)
 
 // on spinlock contention, the side unsuccessfully attempting the lock reverts the lock.
 // if the unlocking side sees the lock incorrectly set, the unlocking side will delay until it has reverted to the correct value, then continue unlocking.
 // there should be a delay of at least about ~10-20 cycles between a lock and unlock, to prevent contention.
-#define SPINLOCK_NOBODY			0x0000	
-#define SPINLOCK_ARM7			0x0001
-#define SPINLOCK_ARM9			0x0002
+#define SPINLOCK_NOBODY			(uint16)(0x0000)
+#define SPINLOCK_ARM7			(uint16)(0x0001)
+#define SPINLOCK_ARM9			(uint16)(0x0002)
 
-#define SPINLOCK_OK		0x0000
-#define SPINLOCK_INUSE	0x0001
-#define SPINLOCK_ERROR	0x0002
+#define SPINLOCK_OK		(uint16)(0x0000)
+#define SPINLOCK_INUSE	(uint16)(0x0001)
+#define SPINLOCK_ERROR	(uint16)(0x0002)
 
 #ifdef ARM7
 #define SPINLOCK_VALUE SPINLOCK_ARM7
